+++ src/main/java/com/ays/user/model/entity/UserEntity.java
@@ -54,10 +54,10 @@ public class UserEntity extends BaseEntity {
     private UserStatus status;
 
     @Column(name = "COUNTRY_CODE")
-    private Integer countryCode;
+    private Long countryCode;
 
     @Column(name = "LINE_NUMBER")
-    private Integer lineNumber;
+    private Long lineNumber;
 
     @Column(name = "ORGANIZATION_ID")
     private String organizationId;
